Webpage Design
This is a rough design for the page. We want to make it clean using a
white page background and simple yet vivid artwork to draw in reader.
The Twitch logos will change from a bright logo to a dim one on mouse
over, the "why extinction" like will simply scroll the user down to our
video. The menu on the top left should also drop down on mouse over
(perhaps calling the it menu at the top and having forums etc. drop down
once moused over) there are links there for the forums and the RSI
application page. The logo in the top middle of the page is set on a
loop there is also a Star Citizen logo featured in the middle of the
page.